Do not break relation on local object diverges if the reference is being set in the remote object.
For instance, if A.id is being set in B.a_id, and we run "a.b = b" and
then "a.id = 1", b.a_id will be changed instead of the relation being
broken.

This is slightly hard to test, as the failure depends on the order
that registered hooks are called. If track_local_changes is called
before break_on_local_diverged, the relation will not be broken since
the remote value will be the same as the local one by the time
break_on_local_diverged is run.