Fix failure of commit() to enforce the timeout, by abstracting the code from raw_execute which enforces the timeout, and reusing it for commit().  Don't bother enforcing the timeout for rollback(), because this shouldn't ever fail due to "database is locked" errors, anyway.
Also, remove the "_started=None" parameter from raw_execute, since it was never
used, and start the timer before the first attempt to perform the execute,
rather than after the first failure.  This should mean that, if pysqlite2
starts enforcing the timeout properly, we don't end up waiting for double the
timeout.