Rolling back last change.
As a side-effect of the introduced change, objects being constructed
could be flushed to the database inside their constructors simply by
accessing a reference which isn't set.  Besides being undesired due
to the additional UPDATE queries that are executed after the CREATE,
this behavior can actually break code if the database expectations
(e.g. NOT NULL) are not entirely satisfied by the time the object
is flushed.

We have ideas for how to implement the desired behavior without
introducing this side effect, but this will require additional
research and time for the implementation.